requirements.txt
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
RPi.GPIO==0.7.0
Werkzeug==1.0.1
